requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.3
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3


fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
google-genai
requests
pandas
python-jose[cryptography]
passlib[bcrypt]
python-multipart